+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.24.0] - 2023-04-07
+
+### Changed
+
+- Node 12 is not supported on this client
+- Version 16 and 18 of nodejs are supported
+- Add Extra information
+- Improve define configuration
+
 ## [3.23.0] - 2023-03-29
 
 ### Changed